+++ src/config/mod.rs
@@ -27,6 +27,8 @@ const DEFAULT_AUXILIARY_ENDPOINT: &str = "http://127.0.0.1:8546";
 const ENV_ORIGIN_CORE_ADDRESS: &str = "MOSAIC_ORIGIN_CORE_ADDRESS";
 const ENV_ORIGIN_VALIDATOR_ADDRESS: &str = "MOSAIC_ORIGIN_VALIDATOR_ADDRESS";
 const ENV_AUXILIARY_VALIDATOR_ADDRESS: &str = "MOSAIC_AUXILIARY_VALIDATOR_ADDRESS";
+const ENV_ORIGIN_BLOCK_STORE_ADDRESS: &str = "MOSAIC_ORIGIN_BLOCK_STORE_ADDRESS";
+const ENV_AUXILIARY_BLOCK_STORE_ADDRESS: &str = "MOSAIC_AUXILIARY_BLOCK_STORE_ADDRESS";
 const ENV_ORIGIN_POLLING_INTERVAL: &str = "MOSAIC_ORIGIN_POLLING_INTERVAL";
 const DEFAULT_ORIGIN_POLLING_INTERVAL: &str = "1";
 const ENV_AUXILIARY_POLLING_INTERVAL: &str = "MOSAIC_AUXILIARY_POLLING_INTERVAL";
@@ -46,6 +48,10 @@ pub struct Config {
     origin_validator_address: Address,
     /// The address that is used to send messages as a validator on auxiliary.
     auxiliary_validator_address: Address,
+    /// The address of origin block store contract.
+    origin_block_store_address: Address,
+    /// The address of auxiliary block store contract.
+    auxiliary_block_store_address: Address,
     origin_polling_interval: Duration,
     auxiliary_polling_interval: Duration,
 }
@@ -104,6 +110,22 @@ impl Config {
                 None => panic!("An auxiliary validator address must be set"),
             };
 
+        let origin_block_store_address =
+            match Self::read_environment_variable(ENV_ORIGIN_BLOCK_STORE_ADDRESS, None) {
+                Some(auxiliary_validator_address) => auxiliary_validator_address
+                    .parse::<Address>()
+                    .expect("The origin block store address cannot be parsed"),
+                None => panic!("An origin block store address must be set"),
+            };
+
+        let auxiliary_block_store_address =
+            match Self::read_environment_variable(ENV_AUXILIARY_BLOCK_STORE_ADDRESS, None) {
+                Some(auxiliary_validator_address) => auxiliary_validator_address
+                    .parse::<Address>()
+                    .expect("The auxiliary block store address cannot be parsed"),
+                None => panic!("An auxiliary block store address must be set"),
+            };
+
         let origin_polling_interval = match Self::read_environment_variable(
             ENV_ORIGIN_POLLING_INTERVAL,
             Some(DEFAULT_ORIGIN_POLLING_INTERVAL),
@@ -140,6 +162,8 @@ impl Config {
             _origin_core_address: origin_core_address,
             origin_validator_address,
             auxiliary_validator_address,
+            origin_block_store_address,
+            auxiliary_block_store_address,
             origin_polling_interval,
             auxiliary_polling_interval,
         }
@@ -200,6 +224,16 @@ impl Config {
         self.auxiliary_validator_address
     }
 
+    /// Returns the address of origin block store.
+    pub fn origin_block_store_address(&self) -> Address {
+        self.origin_block_store_address
+    }
+
+    /// Returns the address of auxiliary block store.
+    pub fn auxiliary_block_store_address(&self) -> Address {
+        self.auxiliary_block_store_address
+    }
+
     pub fn origin_polling_interval(&self) -> Duration {
         self.origin_polling_interval
     }
@@ -239,6 +273,14 @@ mod test {
             ENV_AUXILIARY_VALIDATOR_ADDRESS,
             "1234567890123456789012345678901234567890",
         );
+        env::set_var(
+            ENV_ORIGIN_BLOCK_STORE_ADDRESS,
+            "5678901234123456789012345678901234567890",
+        );
+        env::set_var(
+            ENV_AUXILIARY_BLOCK_STORE_ADDRESS,
+            "5678901234123456789012345678901234567890",
+        );
 
         let config = Config::new();
         assert_eq!(
@@ -277,6 +319,18 @@ mod test {
                 .parse::<Address>()
                 .unwrap()
         );
+        assert_eq!(
+            config.origin_block_store_address(),
+            "5678901234123456789012345678901234567890"
+                .parse::<Address>()
+                .unwrap()
+        );
+        assert_eq!(
+            config.auxiliary_block_store_address(),
+            "5678901234123456789012345678901234567890"
+                .parse::<Address>()
+                .unwrap()
+        );
 
         env::set_var(ENV_ORIGIN_ENDPOINT, "10.0.0.1");
         let config = Config::new();
@@ -305,5 +359,7 @@ mod test {
         env::remove_var(ENV_AUXILIARY_ENDPOINT);
         env::remove_var(ENV_ORIGIN_VALIDATOR_ADDRESS);
         env::remove_var(ENV_AUXILIARY_VALIDATOR_ADDRESS);
+        env::remove_var(ENV_ORIGIN_BLOCK_STORE_ADDRESS);
+        env::remove_var(ENV_AUXILIARY_BLOCK_STORE_ADDRESS);
     }
 }