+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Notable Changes
 
+* Rum now report blocks to origin and auxiliary block store ([#31](https://github.com/OpenSTFoundation/rust-mosaic/pull/31)).
 * Rum now connects to two ethereum nodes ([#27](https://github.com/OpenSTFoundation/rust-mosaic/pull/27)).
 * Docker image is now available at https://hub.docker.com/r/openstfoundation/rust-mosaic/ ([#9](https://github.com/OpenSTFoundation/rust-mosaic/pull/9)).